airflow/pyproject.toml
[project]
name = "airflow-dags"
version = "0.1.0"
description = "This airflow instance is used for pipline Orchestration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.2.0",
    "dbt-bigquery>=1.11.1",
    "google-cloud-bigquery>=3.41.0",
]
